+++ CHANGELOG.md
@@ -14,7 +14,10 @@
 
 #### :nail_care: Polish
 
-Enhance decorator completion. https://github.com/rescript-lang/rescript-vscode/pull/908
+- Enhance decorator completion. https://github.com/rescript-lang/rescript-vscode/pull/908
+- Completion for import attributes in `@module`. https://github.com/rescript-lang/rescript-vscode/pull/913
+- Relax filter for what local files that come up in from and regular string completion in `@module`. https://github.com/rescript-lang/rescript-vscode/pull/918
+- Make from completion trigger for expr hole so we get a nice experience when completing {from: <com>} in `@module`. https://github.com/rescript-lang/rescript-vscode/pull/918
 
 ## 1.38.0